Move the doe save after the algo the iteration
